Move createGallery to shared
